Move Spanned to spanned.rs from lib.rs
